change float to double

+++ AngleComponent/src/main/java/io/asteroidsfx/anglecomponent/AngleComponent.java
@@ -3,7 +3,5 @@ package io.asteroidsfx.anglecomponent;
 import io.asteroidsfx.common.Component;
 
 public class AngleComponent extends Component {
-
-    public float angle;
-
+    public double angle;
 }
