+++ CollisionModule/src/main/java/io/asteroidsfx/collision/CircleColliderComponent.java
@@ -3,7 +3,5 @@ package io.asteroidsfx.collision;
 import io.asteroidsfx.common.Component;
 
 public class CircleColliderComponent extends Component {
-
-    public float radius;
-
+    public double radius;
 }
