+++ Common/src/main/java/io/asteroidsfx/common/System.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public abstract class System {
     public abstract List<Class<? extends Component>> getSignature();
-    public abstract void tick(float dt, List<Entity> entities);
+    public abstract void tick(double dt, List<Entity> entities);
 }
