+++ DragComponent/src/main/java/io/asteroidsfx/dragcomponent/DragComponent.java
@@ -3,7 +3,5 @@ package io.asteroidsfx.dragcomponent;
 import io.asteroidsfx.common.Component;
 
 public class DragComponent extends Component {
-
-    public float drag;
-
+    public double drag;
 }
