+++ MovementSystem/src/main/java/io/asteroidsfx/movementsystem/MovementSystem.java
@@ -18,7 +18,7 @@ public class MovementSystem extends System{
     }
 
     @Override
-    public void tick(float dt, List<Entity> entities) {
+    public void tick(double dt, List<Entity> entities) {
 
         for(Entity entity : entities){
             PositionComponent positionComponent = entity.getComponent(PositionComponent.class);
